fix -Wsign-compare warnings
librtas_src/syscall_calls.c:1381:33: error: comparison of integer
expressions of different signedness: '__uint32_t' {aka 'unsigned int'}
and 'int' [-Werror=sign-compare]
 1381 |         if (be32toh(resp_bytes) > *work_area_bytes) {
      |                                 ^

librtas_src/syscall_rmo.c: In function 'get_phys_region':
librtas_src/syscall_rmo.c:242:42: error: comparison of integer
expressions of different signedness: 'int' and 'uint32_t' {aka
'unsigned int'} [-Werror=sign-compare]
  242 |                 if ((i * WORK_AREA_SIZE) >= kregion->size)
      |                                          ^~

librtas_src/syscall_rmo.c: In function 'release_phys_region':
librtas_src/syscall_rmo.c:291:18: error: comparison of integer
expressions of different signedness: 'uint64_t' {aka 'long unsigned
int'} and 'int' [-Werror=sign-compare]
  291 |         if (bits != ((1 << n_pages) - 1)) {
      |                  ^~

librtasevent_src/get_rtas_event.c: In function 'parse_rtas_event':
librtasevent_src/get_rtas_event.c:411:26: error: comparison of integer
expressions of different signedness: 'uint32_t' {aka 'unsigned int'}
and 'int' [-Werror=sign-compare]
  411 |     if (re->event_length > buflen) {
      |                          ^

librtasevent_src/rtas_srcfru.c: In function 'print_fru_mr_scn':
librtasevent_src/rtas_srcfru.c:518:19: error: comparison of integer
expressions of different signedness: 'int' and 'uint32_t' {aka
'unsigned int'} [-Werror=sign-compare]
  518 |     for (i = 0; i < frumr_num_callouts(fru_mr); i++) {
      |                   ^

This is a step toward enabling a more strict set of warnings by
default.

+++ librtasevent_src/get_rtas_event.c
@@ -408,7 +408,7 @@ parse_rtas_event(char *buf, int buflen)
 
     /* Validate the length of the buffer passed in. */
     re->event_length = re_hdr->ext_log_length + RE_EVENT_HDR_SZ;
-    if (re->event_length > buflen) {
+    if (re->event_length > (uint32_t)buflen) {
         cleanup_rtas_event(re);
         return NULL;
     }
